# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Emby Auto API
  version: 0.22.0
  description: Contract for the Emby Auto API and generated clients.
servers:
  - url: /
    description: Same-origin API served directly or through a reverse proxy
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Create an administrator session
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: The administrator is authenticated.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: emby_auto_session
...
  responses:
    BadRequest:
      description: The request is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Authentication is required or failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    ServiceUnavailable:
      description: A required dependency is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 128
        password:
          type: string
          minLength: 8
          maxLength: 1024
    Session:
      type: object
      additionalProperties: false
      required: [user, expiresAt]
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
        expiresAt:
          type: string
          format: date-time
    AdminUser:
      type: object
      additionalProperties: false
      required: [id, username]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, details, requestId]
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        details:
          type: object
          additionalProperties: true
        requestId:
          type: string
          minLength: 1
...